feature: - convert metrics from  HandlerAspect to Middleware - refactor tests following the feature
